chore: create dom-utils module
